requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
fsspec==2025.9.0
gpt4all==2.8.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
pydub==0.25.1
regex==2025.9.1
requests==2.32.5
SpeechRecognition==3.14.3
sympy==1.14.0
tiktoken==0.11.0
torch==2.8.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
